# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "D-ZO Pay API",
    "description": "Production-grade fintech payment API with multi-provider support, virtual wallets, and admin analytics",
    "version": "1.0.0",
    "contact": {
      "name": "D-ZO Pay Support",
      "email": "support@dzo-pay.com",
      "url": "https://dzo-pay.com/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Local development server"
    },
    {
      "url": "https://api.dzo-pay.com/api/v1",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/wallets": {
      "get": {
        "tags": ["Wallets"],
        "summary": "Get user wallets",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Wallets retrieved" }
        }
      },
      "post": {
        "tags": ["Wallets"],
        "summary": "Create a new wallet",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "currency": { "type": "string", "default": "NGN" },
                  "type": { "type": "string", "enum": ["primary", "savings", "business", "escrow"], "default": "primary" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Wallet created" },
          "400": { "description": "Wallet limit reached" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "adminKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Key"
      }
    }
  }
}